use OwnedHandle

+++ src/platform/windows/tap/mod.rs
@@ -1,4 +1,5 @@
 use crate::platform::windows::{ffi, netsh};
+use std::os::windows::io::{AsRawHandle, OwnedHandle};
 use std::{io, time};
 use windows_sys::Win32::Foundation::HANDLE;
 use windows_sys::Win32::NetworkManagement::Ndis::NET_LUID_LH;
@@ -8,16 +9,18 @@ mod iface;
 
 pub struct TapDevice {
     luid: NET_LUID_LH,
-    handle: HANDLE,
+    handle: OwnedHandle,
     component_id: String,
     index: u32,
+    need_delete: bool,
 }
 unsafe impl Send for TapDevice {}
 unsafe impl Sync for TapDevice {}
 impl Drop for TapDevice {
     fn drop(&mut self) {
-        let _ = ffi::close_handle(self.handle);
-        let _ = iface::delete_interface(&self.component_id, &self.luid);
+        if self.need_delete {
+            let _ = iface::delete_interface(&self.component_id, &self.luid);
+        }
     }
 }
 fn get_version(handle: HANDLE) -> io::Result<[u64; 3]> {
@@ -51,9 +54,8 @@ impl TapDevice {
                     continue;
                 }
                 Ok(handle) => {
-                    if get_version(handle).is_err() {
+                    if get_version(handle.as_raw_handle()).is_err() {
                         std::thread::sleep(time::Duration::from_millis(200));
-                        let _ = ffi::close_handle(handle);
                         continue;
                     }
                     break handle;
@@ -64,7 +66,6 @@ impl TapDevice {
         let index = match ffi::luid_to_index(&luid) {
             Ok(index) => index,
             Err(e) => {
-                let _ = ffi::close_handle(handle);
                 let _ = iface::delete_interface(component_id, &luid);
                 Err(e)?
             }
@@ -74,6 +75,7 @@ impl TapDevice {
             handle,
             index,
             component_id: component_id.to_owned(),
+            need_delete: true,
         })
     }
 
@@ -89,6 +91,7 @@ impl TapDevice {
             luid,
             handle,
             component_id: component_id.to_owned(),
+            need_delete: false,
         })
     }
 
@@ -101,7 +104,13 @@ impl TapDevice {
     /// Retieve the mac of the interface
     pub fn get_mac(&self) -> io::Result<[u8; 6]> {
         let mut mac = [0; 6];
-        ffi::device_io_control(self.handle, TAP_IOCTL_GET_MAC, &(), &mut mac).map(|_| mac)
+        ffi::device_io_control(
+            self.handle.as_raw_handle(),
+            TAP_IOCTL_GET_MAC,
+            &(),
+            &mut mac,
+        )
+        .map(|_| mac)
     }
     pub fn set_mac(&self, _mac: &[u8; 6]) -> io::Result<()> {
         Err(io::Error::from(io::ErrorKind::Unsupported))?
@@ -109,7 +118,7 @@ impl TapDevice {
 
     /// Retrieve the version of the driver
     pub fn get_version(&self) -> io::Result<[u64; 3]> {
-        get_version(self.handle)
+        get_version(self.handle.as_raw_handle())
     }
 
     /// Retieve the mtu of the interface
@@ -149,23 +158,23 @@ impl TapDevice {
         let status: u32 = if status { 1 } else { 0 };
         let mut out_status: u32 = 0;
         ffi::device_io_control(
-            self.handle,
+            self.handle.as_raw_handle(),
             TAP_IOCTL_SET_MEDIA_STATUS,
             &status,
             &mut out_status,
         )
     }
     pub fn try_read(&self, buf: &mut [u8]) -> io::Result<usize> {
-        ffi::try_read_file(self.handle, buf).map(|res| res as _)
+        ffi::try_read_file(self.handle.as_raw_handle(), buf).map(|res| res as _)
     }
     pub fn read(&self, buf: &mut [u8]) -> io::Result<usize> {
-        ffi::read_file(self.handle, buf).map(|res| res as _)
+        ffi::read_file(self.handle.as_raw_handle(), buf).map(|res| res as _)
     }
     pub fn write(&self, buf: &[u8]) -> io::Result<usize> {
-        ffi::write_file(self.handle, buf).map(|res| res as _)
+        ffi::write_file(self.handle.as_raw_handle(), buf).map(|res| res as _)
     }
     pub fn try_write(&self, buf: &[u8]) -> io::Result<usize> {
-        ffi::try_write_file(self.handle, buf).map(|res| res as _)
+        ffi::try_write_file(self.handle.as_raw_handle(), buf).map(|res| res as _)
     }
 }
